chore: create EvaluationStorage on the Data module
