+++ bucketeer/src/main/kotlin/io/bucketeer/sdk/android/internal/evaluation/storage/EvaluationStorage.kt
@@ -14,4 +14,5 @@ internal interface EvaluationStorage {
   fun getBy(featureId: String): Evaluation?
   fun get(): List<Evaluation>
   fun deleteAllAndInsert(evaluations: List<Evaluation>, evaluatedAt: String)
+  fun refreshCache()
 }
