--- bucketeer/src/test/kotlin/io/bucketeer/sdk/android/internal/evaluation/storage/EvaluationStorageImplInitCacheTest.kt
@@ -1,64 +1,0 @@
-package io.bucketeer.sdk.android.internal.evaluation.storage
-
-import android.content.Context
-import androidx.sqlite.db.SupportSQLiteOpenHelper
-import androidx.test.core.app.ApplicationProvider
-import io.bucketeer.sdk.android.internal.Constants
-import io.bucketeer.sdk.android.internal.database.OpenHelperCallback
-import io.bucketeer.sdk.android.internal.database.createDatabase
-import io.bucketeer.sdk.android.internal.di.DataModule
-import io.bucketeer.sdk.android.internal.evaluation.cache.EvaluationSharedPrefs
-import io.bucketeer.sdk.android.internal.evaluation.cache.EvaluationSharedPrefsImpl
-import io.bucketeer.sdk.android.internal.evaluation.cache.MemCache
-import io.bucketeer.sdk.android.internal.evaluation.db.EvaluationSQLDao
-import io.bucketeer.sdk.android.internal.evaluation.db.EvaluationSQLDaoImpl
-import io.bucketeer.sdk.android.internal.model.Evaluation
-import io.bucketeer.sdk.android.mocks.evaluation1
-import io.bucketeer.sdk.android.mocks.evaluation2
-import io.bucketeer.sdk.android.mocks.user1
-import org.junit.After
-import org.junit.Before
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.robolectric.RobolectricTestRunner
-
-@RunWith(RobolectricTestRunner::class)
-class EvaluationStorageImplInitCacheTest {
-  private lateinit var userId: String
-  private lateinit var openHelper: SupportSQLiteOpenHelper
-  private lateinit var evaluationSQLDao: EvaluationSQLDao
-  private lateinit var evaluationSharedPrefs: EvaluationSharedPrefs
-  private lateinit var memCache: MemCache<String, List<Evaluation>>
-
-  @Before
-  fun setUp() {
-    userId = user1.id
-    val moshi = DataModule.createMoshi()
-    val context: Context = ApplicationProvider.getApplicationContext()
-    val sharedPreferences = context.getSharedPreferences(
-      Constants.PREFERENCES_NAME,
-      Context.MODE_PRIVATE,
-    )
-    openHelper = createDatabase(context, OpenHelperCallback.FILE_NAME, sharedPreferences)
-    memCache = MemCache.Builder<String, List<Evaluation>>().build()
-    evaluationSharedPrefs = EvaluationSharedPrefsImpl(sharedPreferences)
-    evaluationSQLDao = EvaluationSQLDaoImpl(openHelper, moshi)
-  }
-
-  @After
-  fun tearDown() {
-    openHelper.close()
-  }
-
-  @Test
-  fun verifyCacheAfterInit() {
-    evaluationSQLDao.put(userId, listOf(evaluation1, evaluation2))
-    val evaluationStorage = EvaluationStorageImpl(
-      userId,
-      evaluationSQLDao,
-      evaluationSharedPrefs,
-      memCache,
-    )
-    assert(evaluationStorage.get() == listOf(evaluation1, evaluation2))
-  }
-}
